requirements.txt
flask
flask_cors
azure-cognitiveservices-speech
python-dotenv
flask-socketio
